Update README.md
From inputting my code in ChatGpt

This is a shell script written for a Raspberry Pi that monitors the temperature of the CPU and GPU and writes the data to a file called "log_temp.txt". The script first sets the variable "cpu" to the temperature of the thermal zone 0, then prints out some ASCII art, and a message displaying the author, and the usage of the script. Then it goes into an infinite loop that runs every 3 seconds, measuring the temperature of the GPU and CPU and writing the data to the log_temp.txt file along with the current time. The script also gives instructions on how to view the contents of the log file and how to end the monitoring.


Any errors ?

I do not see any obvious errors in the script that would prevent it from running. However, it's hard to tell without testing it on a Raspberry Pi and checking for any missing dependencies or permissions issues.

One thing to note is that the script writes the CPU and GPU temperature to the log file but does not write the vcgencmd measure_temp. Also, in the infinite loop the script writes the CPU and GPU temperature in the log file but it does not write the date and time.